Changed float variables to double variables.

+++ src/edu/wit/cs/comp1000/PA2b.java
@@ -21,12 +21,12 @@ public class PA2b {
 
         // Declaring variables a b and c
         System.out.print("Enter a b c: ");
-        float a = scanner.nextFloat();
-        float b = scanner.nextFloat();
-        float c = scanner.nextFloat();
+        double a = scanner.nextFloat();
+        double b = scanner.nextFloat();
+        double c = scanner.nextFloat();
 
         // Finding d, which is the discriminant of the quadratic formula.
-        float d = ((b*b) - 4*a*c);
+        double d = ((b*b) - 4*a*c);
 
         // Checks if the discriminant is greater than, equal too, or less than 0. Depending on this output, the code will either
         // declare the roots are imaginary, declare that there is only one root (variable e), or find that there is two roots that can be calculated.
@@ -35,7 +35,7 @@ public class PA2b {
             System.out.println("Roots: imaginary");
         } else if(d == 0){
         	// If d == 0, it will calculate a special equation to make variable e, which will then be output to the user as one root.
-            float e = (-b/2*a);
+            double e = (-b/2*a);
             System.out.printf("Root: %.2f%n", e);
         } else if(d > 0){
         	// If d is greater than 0, the code will find the roots for both the positive and negative parts of the quadratic formula, and then sort them from smallest to greatest.
